Group sort update; remove fillable
Change group sort to decimal for more precision; remove fillable (not needed as unguarded is used for all models)

+++ database/migrations/2025_06_01_020531_add_sort_order_to_groups_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('groups', function (Blueprint $table) {
-            $table->integer('sort_order')->default(9999);
+            $table->decimal('sort_order', 12, 4)->default(9999);
         });
     }
